Move to accounting/utils/software_plans.py
